Refactor LoginPage to remove error toast for invalid credentials. Update PaymentDetailsDrawer to streamline payment data handling and enhance ReceiptTemplate with back navigation functionality. Modify PropertyPage to clarify parameter naming and improve data fetching logic. Enhance PropertyPaymentHistory with loading states and dynamic payment data display, while updating PropertyOverview for clearer property status representation.

+++ apps/web/services/axiosInstance.ts
@@ -93,9 +93,13 @@ const handleError = async (error: any) => {
     //   }
     //   return Promise.reject(refreshError);
     // }
-    toast.error("Unauthorized: Please log in again.");
-    localStorage.clear();
-    window.location.href = "/login";
+
+    if (!originalRequest.url.includes("login")) {
+      toast.error(data?.message || "Unauthorized: Please log in again.");
+    } else {
+      localStorage.clear();
+      window.location.href = "/login";
+    }
   }
 
   const messages: Record<number, string> = {
